requirements.txt
streamlit==1.38.0
langchain==0.3.1
langchain-community==0.3.1
langchain-google-genai==1.0.8
pypdf==5.0.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
sentence-transformers==3.1.1
google-generativeai==0.7.2